services/device-simulator/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
requests==2.31.0
pydantic==2.5.3
pydantic-settings==2.1.0

